+++ core/src/elements/ons-list-item.js
@@ -113,7 +113,7 @@ export default class ListItemElement extends BaseElement {
    * @attribute tappable
    * @type {Boolean}
    * @description
-   *   [en]Makes the element react to taps.[/en]
+   *   [en]Makes the element react to taps. `prevent-tap` attribute can be added to child elements like buttons or inputs to prevent this effect. `ons-*` elements are ignored by default.[/en]
    *   [ja][/ja]
    */
 
@@ -232,8 +232,27 @@ export default class ListItemElement extends BaseElement {
     }
   }
 
-  _onTouch() {
+  isExcludedTarget(target) {
+
+    if (target.tagName.indexOf('ONS-') === 0) {
+      return true;
+    }
+    if (target.hasAttribute('prevent-tap')) {
+      return true;
+    }
+    if (target.parentElement !== this) {
+      return this.isExcludedTarget(target.parentElement);
+    }
+    return false;
+  }
+
+  _onTouch(e) {
+
     if (this.tapped) {
+      return;
+    }
+
+    if (this !== e.target && this.isExcludedTarget(e.target)) {
       return;
     }