+++ CHANGELOG.md
@@ -11,6 +11,7 @@ dev
  * ons-progress-circular: Fixed [#1860](https://github.com/OnsenUI/OnsenUI/issues/1860).
  * css/polyfills: Fixed [#2266](https://github.com/OnsenUI/OnsenUI/issues/2266).
  * ons-ripple: The effect does not propagate to other `ons-ripple` parents anymore.
+ * ons-carousel: Prevent error state when hidden during the first rendering.
 
  ### Misc