+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGELOG
 dev
 ----
 
+ ### New Features
+
+ * ons-list-item: Skip tappable effect on children with `prevent-tap` attribute or `ons-*` elements.
+
  ### Bug Fixes
 
  * ons-carousel: Fixed [#2260](https://github.com/OnsenUI/OnsenUI/issues/2260).
@@ -17,6 +21,7 @@ dev
 
  * Show warning when Onsen UI is loaded more than once (UMD - ESM).
  * ons-pull-hook: `threshold-height` is not disabled anymore when its value is lower than the height.
+ * angular1: Revised the exposed DOM properties for carousel, modal, navigator, popover, pullHook, splitter, splitterSide, switch and tabbar.
 
 2.8.2
 ----